enhancment: addBooks post table assosiations now will now create a new author is one doesn't exist, but will only accept Genres that are already in the database. the getAllorQueryBooks function will return now turn the only the id, title, author and genre in a single object as opposed to added the Book and Author object as a key value pair inside inside the returned book
enhancment: addBooks post table assosiations now will now create a new
author is one doesn't exist, but will only accept Genres that are
already in the database. the getAllorQueryBooks function will return now
turn the only the id, title, author and genre in a single object as
opposed to added the Book and Author object as a key value pair inside
inside the returned book

ncment: addBooks post table assosiations now will now create a new
author is one doesn't exist, but will only accept Genres that are
already in the database. the getAllorQueryBooks function will return now
turn the only the id, title, author and genre in a single object as
opposed to added the Book and Author object as a key value pair inside
inside the returned book:wq